requirements/base.txt
Django==2.2.10
djangorestframework==3.12.4
psycopg2==2.9.1
djangorestframework-simplejwt==4.8.0
